CreateLocalPlan (feat): integrate with API
