--- app/javascript/react/src/api/localPlan/deleteLocalPlan.js
@@ -1,3 +1,0 @@
-import axios from "axios";
-
-export const deleteLocalPlan = (id) => axios.delete(`local_plan/${id}`);
