--- app/javascript/react/src/api/localPlan/getLocalPlan.js
@@ -1,6 +1,0 @@
-import axios from "axios";
-
-export function getLocalPlan({ queryKey }) {
-  const id = queryKey[1];
-  return axios.get(`/local_plan/${id}`);
-}
