+++ app/views/local_plan_steps/index.html.erb
@@ -1,5 +1,5 @@
 <div>
   <h1 class="font-bold text-4xl">LocalPlanSteps#index</h1>
   <p>Find me in app/views/local_plan_steps/index.html.erb</p>
-  <hello-component name="teste"/>
+  <create-local-plan-component name="teste"/>
 </div>
